requirements.txt
Django=3.0.7
djangorestframework=3.11.0
djoser=2.0.3
psycopg2=2.8.5
matplotlib=3.2.1
Pillow=7.1.2
scipy=1.4.1
mtcnn=0.1.0
tensorflow=2.2.0
keras-vggface=0.6